re-organize some code into helpers file
This partially addresses issue #29 by adding some unit tests, as well.
