+++ vinyldns/helpers.go
@@ -1,0 +1,41 @@
+/*
+Copyright 2018 Comcast Cable Communications Management, LLC
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package vinyldns
+
+import (
+	"strings"
+
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func stringSetToStringSlice(stringSet *schema.Set) []string {
+	ret := []string{}
+	if stringSet == nil {
+		return ret
+	}
+	for _, envVal := range stringSet.List() {
+		ret = append(ret, envVal.(string))
+	}
+	return ret
+}
+
+func parseTwoPartID(id string) (string, string) {
+	parts := strings.Split(id, ":")
+
+	return parts[0], parts[1]
+}
+
+// vinyldns responds 400 to IPv6 addresses represented within `[` `]`
+func removeBrackets(str string) string {
+	return strings.Replace(strings.Replace(str, "[", "", -1), "]", "", -1)
+}
